==> normalization/1_Append_ALL_Carrier_Data.ipynb <==
import pandas as pd
import numpy as np



# Clean AT&T
dfatt=pd.read_csv('../data/att_final.csv')
dfatt.drop(columns=['Extra_Details'],inplace=True)
dfatt['Carrier_Release']=pd.to_datetime(dfatt['Carrier_Release'])
dfatt['Security_Level']=pd.to_datetime(dfatt['Security_Level'])
dfatt=dfatt.sort_values(by=['Carrier_Release'])
dfatt['difference']=dfatt['Carrier_Release']-dfatt['Security_Level']
dfatt['difference']=dfatt['difference'].dt.days
dfatt=dfatt.loc[:,['Manufacture','Phone','Carrier_Release','Security_Level','difference','Android_Level_Guessed',"Baseband","Build_Num","Software_Ver"]]
dfatt=dfatt.astype(str)
dfatt['Build']=dfatt["Baseband"]+" "+dfatt["Build_Num"]+" "+dfatt["Software_Ver"]
dfatt.drop(columns=['Baseband',"Build_Num","Software_Ver"], inplace=True)
dfatt.columns=['Manufacture', 'Model', 'Release_Date', 'Bulletin_Level', 'difference','Android_Level_Guessed',"Build"]

print(dfatt.head())
print(len(dfatt))

# Clean Verizon
dfver=pd.read_csv('../data/VerizonTotal.csv')#,names=['Manufacture','Model', 'Bulletin_Level', 'Release_Date','Details','Android_Level_Guessed'])
dfver=dfver.astype(str)

dfver.Model=dfver.Model.str.strip() #remove special characters
dfver.Model=dfver.Model.str.replace(r"\xa0"," ") #remove special characters

# Calculate the time difference
dfver['Release_Date']=pd.to_datetime(dfver.Release_Date)
dfver['Bulletin_Level']=pd.to_datetime(dfver.Bulletin_Level)
dfver['difference']=dfver['Release_Date']-dfver['Bulletin_Level']
dfver['difference']=dfver['difference'].dt.days
dfver=dfver.loc[:,["Manufacture","Model","Bulletin_Level","Release_Date","difference","Android_Level_Guessed","Version"]]
print(dfver.head())
print(len(dfver))


# Clean Sprint
dfsprnt=pd.read_csv('../data/sprint_final.csv',index_col=False)
dfsprnt=dfsprnt.loc[:,['Manufacture','Model','Release_Date','Bulletin_Level','difference','Description','Android_Level_Guessed','Software_Version']]
dfsprnt=dfsprnt.drop(columns=['Description'])

# Calculate the time differences
dfsprnt['Release_Date']=pd.to_datetime(dfsprnt['Release_Date'])
dfsprnt['Bulletin_Level']=pd.to_datetime(dfsprnt['Bulletin_Level'])
dfsprnt['difference']=dfsprnt['Release_Date']-dfsprnt['Bulletin_Level']
dfsprnt['difference']=dfsprnt['difference'].dt.days

dfsprnt.drop_duplicates(inplace=True)
print(dfsprnt.head())
print(len(dfsprnt))

#Clean T-Mobile
dftmob=pd.read_csv('../data/tmob_final_CLEANED.csv')#,names=['Release_Date','ENHANCEMENTS','Manufacture', 'Model', 'OS' , 'Android_Level_Guessed','Bulletin_Level', 'difference',],index_col=False)
dftmob.drop(columns=['ENHANCEMENTS','OS_version'],inplace=True)
dftmob=dftmob.loc[:,['Manufacture', 'Model', 'Release_Date' , 'Bulletin_Level' , 'difference','Android_Level_Guessed','VERSION_BUILD']]
dftmob['Release_Date']=pd.to_datetime(dftmob['Release_Date'])
dftmob['Bulletin_Level']=pd.to_datetime(dftmob['Bulletin_Level'])
print(dftmob.head())
print(len(dftmob))

### NOTICE: A PROBLEM ARISES FOR PHONE MODEL NAMING AS GALAXY NOTE IS EITHER NOTE {DIGIT} OR NOTE{DIGIT}
# ###################### COMBINE ALL DATA ######################
# drop NaN values
dfatt=dfatt.dropna(subset=['Bulletin_Level'])
dfver=dfver.dropna(subset=['Bulletin_Level'])
print(dfver.columns)

# normalize columns names for future merge
dfatt.columns=['Manufacture', 'Model', 'Release_Date', 'Bulletin_Level','difference','Android_Level_Guessed',"build"]
dfatt['Carrier']="AT&T"

dfver.columns=['Manufacture', 'Model','Bulletin_Level','Release_Date', 'difference','Android_Level_Guessed',"build"]
dfver['Carrier']="Verizon"

dftmob.columns=['Manufacture', 'Model', 'Release_Date', 'Bulletin_Level','difference','Android_Level_Guessed',"build"]
dftmob['Carrier']="TMobile"

dfsprnt.columns=['Manufacture', 'Model', 'Release_Date', 'Bulletin_Level','difference','Android_Level_Guessed',"build"]
dfsprnt['Carrier']="Sprint"


# append all together
finalDF=dfatt
finalDF=finalDF.append(dfver)
finalDF=finalDF.append(dftmob)
finalDF=finalDF.append(dfsprnt)

# shift carrier columns
columns=list(finalDF)
columns.insert(0, columns.pop(columns.index('Carrier')))
finalDF= finalDF.loc[:, columns]

# sort by carrier, manufacture, model, bulletin level
print(len(finalDF))
print(finalDF.head())


#Remove outliers
finalDF.difference=finalDF.difference.astype('double')

# Save the remaining dataframe!!
finalDF.to_csv('../data/allCarrierData_merged.csv',index=False)


# All higher-level bulletins
finalDF['checker']=finalDF.Bulletin_Level.apply(lambda x: str(x)[-2:])
print(finalDF[finalDF.checker!="01"])

#print(finalDF.Model.unique())
print(finalDF.Model.unique())
