==> normalization/2_AddReleaseYear.ipynb <==
import pandas as pd
import numpy as np

dfCar=pd.read_csv('../data/allCarrierData_merged.csv')
dfRelease=pd.read_csv('../data/RecentlyReleasedPhones.csv')

dfCar.Manufacture=dfCar.Manufacture.astype('str')
dfCar.Model=dfCar.Model.str.strip()

# First, adjust time of release... only consider August 2015 and later
print(len(dfCar))
dfCar.Release_Date=pd.to_datetime(dfCar.Release_Date)

dfCar=dfCar[dfCar.Release_Date>=pd.to_datetime("2015-08-01")]
dfCar=dfCar[dfCar.Release_Date<pd.to_datetime("2020-01-01")]
print(len(dfCar))

# Adjust phone names as needed!
dfCar.Manufacture=dfCar.Manufacture.str.replace("^Other$","ASUS")
dfCar.Manufacture=dfCar.Manufacture.str.replace("^Blackberry$","BlackBerry")
dfCar.Manufacture=dfCar.Manufacture.str.replace("^Samgung$","Samsung")
dfCar.Manufacture=dfCar.Manufacture.str.replace("^kyocera$","Kyocera")
dfCar.Manufacture=dfCar.Manufacture.str.replace("^lg$","LG")
dfCar.Manufacture=dfCar.Manufacture.str.replace("^palm$","Palm")

# Any galaxy note phones with digits should be note#
dfCar.Model=dfCar.Model.str.strip()
if len(dfCar[dfCar.Model.str.contains("Note ")])>0:
    for i in range(10):
        dfCar.Model=dfCar.Model.str.replace("Note "+str(i),"Note"+str(i))

# S6 Edge Plus AND Galaxy S6 Edge + to Galaxy S6 Edge+ 
if len(dfCar[dfCar.Model.str.contains("S6 Edge Plus")])>0:
    for i in range(10):
        dfCar.Model=dfCar.Model.str.replace("S6 Edge Plus","Galaxy S6 Edge+")
        dfCar.Model=dfCar.Model.str.replace("Galaxy S6 Edge +","Galaxy S6 Edge+")
        
# Normalize Motorola phones as needed
if len(dfCar.Model.str.contains("moto "))>0:
    dfCar.Model=dfCar.Model.str.replace("moto ","Moto ")

dfCar.Model=dfCar.Model.str.replace("^Moto Z2 Force Edition","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^Moto z2 force edition","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^moto z2 force edition","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^Motorola Moto Z2 Force Edition$","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^Motorola Moto X$","Moto X")
dfCar.Model=dfCar.Model.str.replace("^Motorola Nexus 6$","Nexus 6")
dfCar.Model=dfCar.Model.str.replace("^g6$","Moto g6")

# Prepend LG as needed
dfCar.Model=dfCar.Model.str.replace(r"^LG G6 \(H871$","G6")
dfCar.Model=dfCar.Model.str.replace(r"LG  H871S\)$","G6")
indices=dfCar[dfCar.Model=="G6"].index
dfCar.at[indices,'Manufacture']='LG'

dfCar.Model=dfCar.Model.str.replace(r"^Google Pixel 4$","Pixel 4")
indices=dfCar[dfCar.Model=="Pixel 4"].index
dfCar.at[indices,'Manufacture']='Google'

dfCar.Model=dfCar.Model.str.replace(r"^Google  Pixel 4 XL$","Pixel 4 XL")
indices=dfCar[dfCar.Model=="Pixel 4 XL"].index
dfCar.at[indices,'Manufacture']='Google'

dfCar.Model=dfCar.Model.str.replace("^G7 ThinQ","LG G7 ThinQ")
dfCar.Model=dfCar.Model.str.replace("^LG V20$","V20")
dfCar.Model=dfCar.Model.str.replace("^LG G2$","G2")
dfCar.Model=dfCar.Model.str.replace("^LG G3$","G3")
dfCar.Model=dfCar.Model.str.replace("^LG G4$","G4")
dfCar.Model=dfCar.Model.str.replace("^LG G5$","G5")
dfCar.Model=dfCar.Model.str.replace("^LG G6$","G6")
dfCar.Model=dfCar.Model.str.replace("^G8","G8 ThinQ")
dfCar.Model=dfCar.Model.str.replace("G8 ThinQ ThinQ","G8 ThinQ")
dfCar.Model=dfCar.Model.str.replace("^LG K10$","K10")
dfCar.Model=dfCar.Model.str.replace("^LG K20$","K20")
dfCar.Model=dfCar.Model.str.replace("LG Phoenix 3","K20") # AT&T specific...
dfCar.Model=dfCar.Model.str.replace("LG Phoenix 2","LG K8") # AT&T specific...
dfCar.Model=dfCar.Model.str.replace("Stylo 2 PLUS","LG Stylo 2 Plus")
dfCar.Model=dfCar.Model.str.replace("^Stylo 3$","LG Stylo 3")
dfCar.Model=dfCar.Model.str.replace("^Stylo 3 Plus$","LG Stylo 3 Plus")
dfCar.Model=dfCar.Model.str.replace("^Aristo$","LG Aristo")
dfCar.Model=dfCar.Model.str.replace("^Aristo 2$","LG Aristo 2")
dfCar.Model=dfCar.Model.str.replace("Aristo 2 PLUS","LG Aristo 2 PLUS")
dfCar.Model=dfCar.Model.str.replace("Leon LTE","Leon")
dfCar.Model=dfCar.Model.str.replace("^Q7\+$","LG Q7+")


# Other phones
dfCar.Model=dfCar.Model.str.replace("^ZTE Maven","Maven")
dfCar.Model=dfCar.Model.str.replace("^ZTE Maven 2","Maven 2")
dfCar.Model=dfCar.Model.str.replace("^Nexus 7 2013 LTE","Google Nexus 7 (2013)")
dfCar.Model=dfCar.Model.str.replace("^J3 Prime","Galaxy J3 Prime")  
dfCar.Model=dfCar.Model.str.replace("^One$","HTC One")
dfCar.Model=dfCar.Model.str.replace("^One LTE$","HTC One")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Note10\.1 2014 Edition$","Galaxy Note 10.1 (2014)")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J7 Star$","Galaxy J7 (2018)")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J3 Star$","Galaxy J3 (2018)")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J3 V \(3rd Gen\.\)$","Galaxy J3 (2018)")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J7 V \(2nd Gen\.\)$","Galaxy J7 (2018)")
dfCar.Model=dfCar.Model.str.replace("^HTC One M9$","One M9")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy Tab A \(8\.0\)$","Galaxy Tab A 8.0 (2018)")
dfCar.Model=dfCar.Model.str.replace("^Essential Phone$","Essential PH-1")
dfCar.Model=dfCar.Model.str.replace("^Xperia Z1S 4G LTE$","Sony Xperia Z1S")
dfCar.Model=dfCar.Model.str.replace("^6T$","OnePlus 6T")
dfCar.Model=dfCar.Model.str.replace("^Kyocera DuraForce Pro$","DuraForce Pro")
dfCar.Model=dfCar.Model.str.replace("^Tab A \(2018\)$","Galaxy Tab A 8.0 (2018)")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy Tab A \(10\.5\)$","Galaxy Tab A (2018, 10.5)")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J3 Eclipse$","Galaxy J3 Emerge")
dfCar.Model=dfCar.Model.str.replace("^Galaxy J3 Prime$","Galaxy J3 Emerge")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Express Prime 2$","Galaxy J3 Emerge")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Note10\+ 5G$","Galaxy Note10+")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Note10 5G$","Galaxy Note10")
dfCar.Model=dfCar.Model.str.replace("^KEYone$","BlackBerry KEYone")
dfCar.Model=dfCar.Model.str.replace("^HTC 10$","10")

indices=dfCar[dfCar.Model.str.contains("Galaxy")].index
dfCar.at[indices,'Manufacture']='Samsung'

indices=dfCar[dfCar.Manufacture.str.contains("Pointing out that this")].index
dfCar.at[indices,'Manufacture']='Samsung'

# Fix capitalization
dfCar.Model=dfCar.Model.str.replace("Nexus 5X","Nexus 5x")
dfCar.Model=dfCar.Model.str.replace("Nexus 6P","Nexus 6p")
dfCar.Model=dfCar.Model.str.replace("^Moto G6 Play$","Moto g6 play")
dfCar.Model=dfCar.Model.str.replace("^Moto G7 power$","Moto g7 power")
dfCar.Model=dfCar.Model.str.replace("^Moto E4$","Moto e4")
dfCar.Model=dfCar.Model.str.replace("^Moto E4 Plus$","Moto e4 plus")
dfCar.Model=dfCar.Model.str.replace("^Moto E5 Play$","Moto e5 play")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S10 5G$","Galaxy S10")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S7 edge$","Galaxy S7 Edge")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S6 edge$","Galaxy S6 Edge")
dfCar.Model=dfCar.Model.str.replace("^Galaxy GRAND Prime$","Galaxy Grand Prime")
dfCar.Model=dfCar.Model.str.replace("^6$","Nexus 6")

indices=dfCar[dfCar.Model=="Nexus 6"].index #all Nexus 6 devices should be Motorola (this was designed incoordination with Google)
dfCar.at[indices,'Manufacture']='Motorola'

indices=dfCar[dfCar.Model=="Nexus 6x"].index #all Nexus 6x devices should be LG (this was designed incoordination with Google)
dfCar.at[indices,'Manufacture']='LG'

indices=dfCar[dfCar.Model=="Nexus 6p"].index #all Nexus 6 devices should be Huawei (this was designed incoordination with Google)
dfCar.at[indices,'Manufacture']='Huawei'


# Some patches apply to multiple phones. Must separate these patches as needed...
print("######## Need to split these models into different patches ########")
if len(dfCar.Model.str.contains(" and "))>0:
    print(dfCar[dfCar.Model.str.contains(" and ")].Model)
    
if len(dfCar.Model.str.contains("/"))>0:
    dfCar.Model=dfCar.Model.str.replace("Phoenix Plus / ","") # Remove odd extra 
    dfCar_toAppend=dfCar[dfCar.Model.str.contains("/V30+")].copy()
    dfCar.Model=dfCar.Model.str.replace("/V30+","")
    dfCar_toAppend.Model=dfCar_toAppend.Model.str.replace("V30/","")
    print("Number of updates for multiple models")
    print(len(dfCar_toAppend))
    print(dfCar_toAppend.Carrier.unique())
    dfCar=dfCar.append(dfCar_toAppend)
    print(dfCar[dfCar.Model.str.contains("/")].Model)
    
    
# Fix any remaining issues
dfCar.Model=dfCar.Model.str.replace("( )+"," ")
dfCar.Model=dfCar.Model.str.replace("^e4$","Moto e4")
dfCar.Model=dfCar.Model.str.replace("^e4 plus$","Moto e4 plus")
dfCar.Model=dfCar.Model.str.replace("^e4 gold$","Moto e4 gold")
dfCar.Model=dfCar.Model.str.replace("^e5 go$","Moto e5 go")
dfCar.Model=dfCar.Model.str.replace("^e5 play$","Moto e5 play")
dfCar.Model=dfCar.Model.str.replace("^g6 play$","Moto g6 play")
dfCar.Model=dfCar.Model.str.replace("^g7 power$","Moto g7 power")
dfCar.Model=dfCar.Model.str.replace("^z2 force edition","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^Moto z2 force","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^z3","Moto z3")
dfCar.Model=dfCar.Model.str.replace("^z4","Moto z4")
dfCar.Model=dfCar.Model.str.replace("^2 V","Nokia 2 V")
dfCar.Model=dfCar.Model.str.replace("^PRO 2","DuraForce PRO 2")
dfCar.Model=dfCar.Model.str.replace("^PRO with Sapphire Shield","DuraForce PRO with Sapphire Shield")
dfCar.Model=dfCar.Model.str.replace("^Kyocera DuraForce PRO 2","DuraForce PRO 2")
dfCar.Model=dfCar.Model.str.replace("^Kyocera DuraForce PRO with Sapphire Shield","DuraForce PRO with Sapphire Shield")
dfCar.Model=dfCar.Model.str.replace("^Stylo 4\+","LG Stylo 4+")
dfCar.Model=dfCar.Model.str.replace("^Stylo 5","LG Stylo 5")
dfCar.Model=dfCar.Model.str.replace("^Stylo 2","LG Stylo 2")
dfCar.Model=dfCar.Model.str.replace("^Stylo 4","LG Stylo 4")
dfCar.Model=dfCar.Model.str.replace("^Stylo 2 V","LG Stylo 2 V")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Tab E \(8.0\)","Galaxy Tab E 8.0")
dfCar.Model=dfCar.Model.str.replace("^Galaxy Tab E$","Galaxy Tab E 8.0")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy S6 edge [+]+","Galaxy S6 Edge+")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy S6 Edge[+]{2}","Galaxy S6 Edge+")
dfCar.Model=dfCar.Model.str.replace("^z2 play","Moto z2 play")
dfCar.Model=dfCar.Model.str.replace("^Motorola Moto e5 play","Moto e5 play")
dfCar.Model=dfCar.Model.str.replace("^Motorola Moto G6 Play","Moto G6 Play")
dfCar.Model=dfCar.Model.str.replace("^Palm Palm","Palm")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S 4","Galaxy S4")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S 5","Galaxy S5")
dfCar.Model=dfCar.Model.str.replace("^Galaxy S 4 mini","Galaxy S4 mini")
dfCar.Model=dfCar.Model.str.replace("^E5 Play","Moto e5 play")
dfCar.Model=dfCar.Model.str.replace(r"^HTC One \(M8\)$","One M8") 
dfCar.Model=dfCar.Model.str.replace("^One \(M8\)$","One M8") 
dfCar.Model=dfCar.Model.str.replace("^DROID MINI","Droid Mini")
dfCar.Model=dfCar.Model.str.replace("^Droid TURBO","Droid Turbo")
dfCar.Model=dfCar.Model.str.replace("^DROID TURBO 2","Droid Turbo 2")
dfCar.Model=dfCar.Model.str.replace("^Phone","Palm")
dfCar.Model=dfCar.Model.str.replace("^E 4th Generation","Moto E 4th Generation")
dfCar.Model=dfCar.Model.str.replace(" \(J727A\)","")
dfCar.Model=dfCar.Model.str.replace('Galaxy Note\x01',"Galaxy Note")
dfCar.Model=dfCar.Model.str.replace("^Z3 play","Moto Z3 play")
dfCar.Model=dfCar.Model.str.replace("^Z3 Play","Moto Z3 play")
dfCar.Model=dfCar.Model.str.replace("^Z2 Force","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace("^G7 Play","Moto G7 play")
dfCar.Model=dfCar.Model.str.replace("^HTC One A9","One A9")
dfCar.Model=dfCar.Model.str.replace("^LG G8X ThinQ","G8 ThinQ")
dfCar.Model=dfCar.Model.str.replace("^e6$","Moto e6")
dfCar.Model=dfCar.Model.str.replace("^LG V30$","V30")
dfCar.Model=dfCar.Model.str.replace("^G6$","LG G6")
dfCar.Model=dfCar.Model.str.replace("^GizmoTablet$","Gizmo Tablet")
dfCar.Model=dfCar.Model.str.replace("^LG G8$","LG G8 ThinQ")

dfCar.Manufacture=dfCar.Manufacture.str.replace("^DuraForce$","Kyocera")
    
dfCar.Model=dfCar.Model.str.replace("^HYDROGEN ONE$","RED Hydrogen One")
dfCar.Model=dfCar.Model.str.replace("^Orbic Wonder$","Wonder")
dfCar.Model=dfCar.Model.str.replace("^Tab A 8\.0$","Galaxy Tab A 8.0 (2018)")
dfCar.Model=dfCar.Model.str.replace("^HTC Desire 626$","Desire 626")
dfCar.Model=dfCar.Model.str.replace(r"^Slate 8$","Slate 8 Tablet")
dfCar.Model=dfCar.Model.str.replace(r"^Moto Z2 Force","Moto Z (2) Force")
dfCar.Model=dfCar.Model.str.replace(r"^GS7 Edge Special Edition$","Galaxy S7 Edge Special Edition")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy S7 Edge special edition$","Galaxy S7 Edge Special Edition")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy Express Prime 3$","Galaxy Express Prime 3")
dfCar.Model=dfCar.Model.str.replace(r"^Moto G Play$","Moto G4 Play")
dfCar.Model=dfCar.Model.str.replace(r"^Tab A 8\.0$","Galaxy Tab A 8.0 (2018)")
dfCar.Model=dfCar.Model.str.replace(r"^Galaxy Tab A 8\.0$","Galaxy Tab A 8.0 (2018)")
dfCar.Model=dfCar.Model.str.replace(r"^LG G6$","G6")
dfCar.Model=dfCar.Model.str.replace(r"^Alcatel IDEAL$","Alcatel OneTouch IDEAL")
dfCar.Model=dfCar.Model.str.replace(r"Moto z2 play^$","Moto Z (2) Play")
dfCar.Model=dfCar.Model.str.replace(r"^LG G8 ThinQ$","G8 ThinQ")
dfCar.Model=dfCar.Model.str.replace(r"^8 Tablet$","Slate 8 Tablet")
dfCar.Model=dfCar.Model.str.replace(r"^Moto z2 play$","Moto Z (2) play")
dfCar.Model=dfCar.Model.str.replace(r"^G6 Play$","moto g6 play")
dfCar.Model=dfCar.Model.str.replace(r"^E5 Plus$","moto e5 plus")
dfCar.Model=dfCar.Model.str.replace(r"^E4 Plus$","moto e4 plus")




if len(dfCar.Model.str.contains(" and "))>0:
    toAppend=dfCar[dfCar.Model.str.contains(" and ")]
    print("Total number of updates for multiple models")
    print(len(toAppend))
    print(toAppend.Carrier.unique())
    toAppend.reset_index(drop=False, inplace=False)
    toAppend.model="Alcatel CAMEOX"
    dfCar.append(toAppend)
    dfCar.reset_index(drop=False, inplace=False)
    dfCar.Model=dfCar.Model.str.replace(r"^Alcatel IdealXCITE and CAMEOX$","Alcatel IdealXCITE")
#     print(dfCar[dfCar.Model.str.contains("Alcatel")].Model) 


dfCar.Model=dfCar.Model.str.strip()
dfCar.Manufacture=dfCar.Manufacture.str.strip()


######### Start Appending Release Years ######### 
dfCar['modelMerge']=dfCar.Model.str.strip()
dfCar['modelMerge']=dfCar.modelMerge.str.rstrip(' ')
dfCar['modelMerge']=dfCar.modelMerge.str.lower()
print(dfCar.head())

print("--------------------")
print(dfRelease.columns)
dfRelease.drop(columns=['Released OS','Current Supported OS','Reference Source','Android One'],inplace=True)
dfRelease.columns=['Manufacture','Model','Phone_Release']
dfRelease['Phone_Release']=dfRelease['Phone_Release'].str.replace(r'^([0-9]{4})/([0-9])$',r"\1/0\2") # normalize dates
dfRelease['modelMerge']=dfRelease.Model.str.lstrip(' ')
dfRelease['modelMerge']=dfRelease.modelMerge.str.rstrip(' ')
dfRelease['modelMerge']=dfRelease.modelMerge.str.lower()
dfRelease.drop(columns=['Manufacture','Model'],inplace=True)
print(dfRelease.head())

test=pd.merge(dfCar, dfRelease, on=['modelMerge'],how='left')
test.Phone_Release=test.Phone_Release.astype('str')
test['Year_Phone_Release']=test.Phone_Release.copy()
test['Year_Phone_Release']=test.Year_Phone_Release.apply(lambda x: x[:4])
print(test.head())

print(len(dfCar))
print(len(test))
print(len(dfCar)-len(test))
print(len(test)/len(dfCar))

noGood=[]
for index,value in test.iterrows():
    if value.Phone_Release=='nan':
        noGood.append(value.values)
        #print(value.values)

print(len(noGood))

noGoodDF=pd.DataFrame(noGood,columns=['Carrier','Manufacture','Model','Release_Date','Security_Patch','difference','Android_Level_Guessed', 'build','modelMerge','Phone_Release','Year_Phone_Release'])
uniqPhone=noGoodDF.modelMerge.unique()
print(len(uniqPhone))
print(noGoodDF.head())

noGoodCounts = noGoodDF.groupby(['Manufacture','Model'])['Release_Date'].nunique()
noGoodCounts=noGoodCounts.sort_values(ascending=False)

print("Unique phones: "+str(len(noGoodCounts)))
# print(noGoodCounts[noGoodCounts==7]) #.iloc[48:-10]
# print(noGoodCounts)


for index,value in noGoodCounts.iteritems():
    print(str(index[0])+";"+str(index[1])+";"+str(value))

# print("-----------------------------------")
# print(test[test.Manufacture.str.contains("Nexus")])
#print(test[test.Manufacture=="HTC"])

# print(test[test.Model=="6"])
# print(test[test.Model.str.contains("2nd")])#.Model.unique())



# ASSUMPTION
    # Stylo 4 is LG Q Stylo 4
    # Moto X is 1st generation
    # LG Aristo: https://pc-tablet.com/lg-aristo-price-specs-features-and-release-date-in-the-us/

# print(test[test.Model=="6T"])

###### Result to save
finalDF=pd.merge(dfCar, dfRelease, on=['modelMerge'],how='inner')
finalDF.Phone_Release=finalDF.Phone_Release.astype('str')
finalDF['Year_Phone_Release']=finalDF.Phone_Release.copy()
finalDF['Year_Phone_Release']=finalDF.Year_Phone_Release.apply(lambda x: x[:4])

finalDF['Month_Phone_Release']=finalDF.Phone_Release.copy()
finalDF['Month_Phone_Release']=finalDF.Month_Phone_Release.apply(lambda x: x[-2:])
print(len(finalDF))
print(len(finalDF)/len(dfCar)*100)
print(finalDF.head())

finalDF.sort_values(by=['Carrier','Manufacture','Model','Release_Date'],inplace=True)
finalDF.drop(columns=['modelMerge','Phone_Release'],inplace=True)



#dfCar=pd.read_csv('../Carrier_Latency/allCarrierData_cleaned.csv')
print("Original File: ")
print(len(dfCar))
print()
print("------------------------------------------------")
print("Cleaned File: ")
print(len(finalDF))
print()
print("AT&T")

print("Sprint")

print("Tmobile")

print("Verizon")


print("Percentage of Usable Updates: ")
percent=(len(finalDF)/len(dfCar))*100
print(percent)

print(len(finalDF))
print(finalDF.Carrier.unique())
print("AT&T")
print(len(finalDF[finalDF.Carrier=='AT&T']))
print("Sprint")
print(len(finalDF[finalDF.Carrier=='Sprint']))
print("Tmobile")
print(len(finalDF[finalDF.Carrier=='TMobile']))
print("Verizon")
print(len(finalDF[finalDF.Carrier=='Verizon']))

finalDF.to_csv('../data/allCarrierData_merged_years.csv',index=False)

print(len(dfCar.Model.unique()))
print(len(dfCar.Manufacture.unique()))
print("--------------------------------")
print(len(finalDF.Model.unique()))
print(len(finalDF.Manufacture.unique()))


# print(dfCar[dfCar.Manufacture=="Pointing out that this OS update is specific to the S7 Special Edition only. Samsung"])
print(dfCar.Manufacture.unique())


# Look at phone release vs update release. Any updates reported prior?

# print(finalDF.head())


def monthAdd(mon):
    if mon==12: return "01"
    elif mon>=9: return str(mon+1)
    else: return "0"+str(mon+1)
    
finalDF['Month_Phone_Release_add']=finalDF['Month_Phone_Release'].apply(lambda x: monthAdd(int(x)))

finalDF['PhoneRelease']=finalDF['Year_Phone_Release'].astype(str)+"-"+finalDF['Month_Phone_Release_add'].astype(str)+"-01"
finalDF['PhoneRelease']=pd.to_datetime(finalDF['PhoneRelease'])
finalDF['Release_Date']=pd.to_datetime(finalDF['Release_Date'])


print()

hm=finalDF[finalDF['Release_Date']<finalDF['PhoneRelease']]

print(len(hm))
print(hm.Model.unique())

for mod in hm.Model.unique():
    tmps=hm[hm.Model==mod]
    tmps=tmps.loc[:,['Carrier','Manufacture','Model','Release_Date','PhoneRelease']]
    print(len(tmps))
    print(tmps.head())
    
    print("--------------------------------------------------")
    print("--------------------------------------------------")



# mavin 2 -- had 3 updates
# AT&T     Samsung  Galaxy J3 (2017) -- has 1 updates
# Thanks Sprint... they renamed Illumina Coolpad Legacy Go renamed a phone model
# Idek anymore...
