==> normalization/3_Carrier_normalize_and_timeFilter.ipynb <==
## Important Note
Some manual data manipulation occurred on this dataset! If you're attempting to data clean from start (initial scrape) to finish, there may be some discrepancies.

#Import libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt 

carrierDF=pd.read_csv('../data/allCarrierData_merged_years.csv')
print(len(carrierDF))

# Drop duplicates
carrierDF.drop_duplicates(keep='first', inplace=True ) # drop whole duplicate rows
#carrierDF[carrierDF.duplicated(keep=False)]



carrierDF['checker']=carrierDF.Bulletin_Level.apply(lambda x: str(x)[:10])
carrierDF['checker']=carrierDF.checker.apply(lambda x: str(x)[-2:])
print(carrierDF[carrierDF.checker!="01"])

######## Filter down to approprate timeline ########
print(len(carrierDF))

carrierDF.Bulletin_Level=carrierDF.Bulletin_Level.astype(str)
carrierDF.Bulletin_Level=carrierDF.Bulletin_Level.str.replace(" [0:]*","")
carrierDF.Bulletin_Level=carrierDF.Bulletin_Level.apply(lambda x: str(x[:-2])+"01") # convert all bulletins to first day

# setting the bad bulletins
#badBulletins=['2019-09-01', '2019-08-01', '2019-07-01']

# for badBul in badBulletins: # filtering...
#     carrierDF=carrierDF[carrierDF.Bulletin_Level!=badBul]
print(len(carrierDF))

carrierDF[carrierDF.duplicated(subset=['Carrier','Manufacture','Model','Bulletin_Level'],keep=False)]

# Remove duplicate carrier, manufacture, phone, bulletin pairings. Keep the earliest release date
carrierDF=carrierDF.sort_values(by=['Carrier','Manufacture','Model','Bulletin_Level','difference'])# Sort to properly drop duplicates if bulletins were normalized to the same (e.g., level 1 and 5 were both patched thus keep the first date)
carrierDF.drop_duplicates(subset=['Carrier','Manufacture','Model','Bulletin_Level'],keep='first', inplace=True ) # drop the duplicates and keep the lowest difference
print(len(carrierDF))

for car in carrierDF.Carrier.unique():
    print(car)
    print(len(carrierDF[carrierDF.Carrier==car]))

# errors in parsing verizon...
carrierDF=carrierDF.sort_values(by=['Carrier','Manufacture','Model','Release_Date','Bulletin_Level'])
carrierDF.drop_duplicates(subset=['Carrier','Manufacture','Model','Release_Date','Bulletin_Level'],keep='first', inplace=True ) 

print(len(carrierDF))

# Recalculate difference!
carrierDF.Bulletin_Level=pd.to_datetime(carrierDF.Bulletin_Level)
carrierDF.Release_Date=pd.to_datetime(carrierDF.Release_Date)

newDif=carrierDF.Release_Date-carrierDF.Bulletin_Level
newDif=newDif.apply(lambda x: x.days)
carrierDF.difference=newDif.copy()
print(carrierDF.head())

# Remove outliers!!
index=carrierDF[((carrierDF.Model=='Galaxy A6') & (carrierDF.difference==-312))].index
carrierDF.at[index,'Release_Date']=pd.to_datetime("2019-05-24")

index=carrierDF[((carrierDF.Model=='Galaxy Tab A 8.0 (2018)') & (carrierDF.difference==-71))].index
carrierDF.at[index,'Release_Date']=pd.to_datetime("2019-06-22")

#The following is a use case from using the assumption of the "same" year eventhough Jan starts a new year
index=carrierDF[((carrierDF.Model=='Galaxy Note9') & (carrierDF.difference==-329))].index
carrierDF.at[index,'Bulletin_Level']=pd.to_datetime("2018-12-01")


# Recalculate readjusted outliers
carrierDF.Bulletin_Level=pd.to_datetime(carrierDF.Bulletin_Level)
carrierDF.Release_Date=pd.to_datetime(carrierDF.Release_Date)

newDif=carrierDF.Release_Date-carrierDF.Bulletin_Level
newDif=newDif.apply(lambda x: x.days)
carrierDF.difference=newDif.copy()

carrierDF=carrierDF[carrierDF.difference>=-10]
carrierDF=carrierDF[carrierDF.difference<=150]


# Now change T-Mobile manufacturer to Revvl (T-Mobile)
carrierDF.Manufacture=carrierDF.Manufacture.str.replace("T-Mobile","Revvl(T-Mobile)")
carrierDF.Carrier=carrierDF.Carrier.str.replace("TMobile","T-Mobile")


print(len(carrierDF))
print(len(carrierDF.Model.unique()))
print(len(carrierDF.Manufacture.unique()))
print()
for car in carrierDF.Carrier.unique():
    print(car)
    print(len(carrierDF[carrierDF.Carrier==car]))

carrierDF.sort_values(by=['Carrier','Manufacture','Model','Release_Date'],inplace=True)
carrierDF.to_csv('../data/allCarrierData_final.csv',index=False)

print(carrierDF.head())

print(carrierDF.Manufacture.unique())

# print(carrierDF[carrierDF.Model.str.contains("Gizmo")])
